requirements.txt
discord.py
mysql.connector
discord-py-slash-command